Mrtr tests updates (#3)
* remove duplicates in index, rename test cases to be consistent

* add negative tests

* refactor into everything-server and update negative tests

* fix conformance tests
