+++ src/seps/sep-2322.yaml
@@ -15,22 +15,22 @@ requirements:
     text: 'Servers MUST NOT send InputRequiredResult responses on any other client requests.'
 
   # ── Server Requirements (Basic Workflow) ───────────────────────────────────
-  - check: input-required-result-multiple-inputs-incomplete
+  - check: sep-2322-multiple-inputs-incomplete
     text: 'inputRequests keys are server assigned identifiers and MUST be unique within the scope of the request.'
 
-  - check: input-required-result-elicitation-incomplete
+  - check: sep-2322-elicitation-incomplete
     text: 'inputRequests values are request objects that MUST be one of ElicitRequest, CreateMessageRequest, or ListRootsRequest (elicitation variant)'
 
-  - check: input-required-result-sampling-incomplete
+  - check: sep-2322-sampling-incomplete
     text: 'inputRequests values are request objects that MUST be one of ElicitRequest, CreateMessageRequest, or ListRootsRequest (sampling variant)'
 
-  - check: input-required-result-list-roots-incomplete
+  - check: sep-2322-list-roots-incomplete
     text: 'inputRequests values are request objects that MUST be one of ElicitRequest, CreateMessageRequest, or ListRootsRequest (list roots variant)'
 
   - check: sep-2322-reject-tampered-state
     text: 'If requestState influences authorization, resource access, or business logic, servers MUST protect its integrity and MUST reject state that fails verification.'
 
-  - check: input-required-result-request-state-incomplete
+  - check: sep-2322-request-state-incomplete
     text: 'Servers MUST include at least one of inputRequests or requestState in every InputRequiredResult response.'
 
   - check: sep-2322-respect-client-capabilities
@@ -50,36 +50,36 @@ requirements:
     text: "Both the inputRequests and requestState fields affect only the client's retry of the original request. They MUST NOT be used for any other request that the client may be sending in parallel."
 
   # ── Ephemeral Workflow: Complete Result After Retry ─────────────────────────
-  - check: input-required-result-elicitation-complete
+  - check: sep-2322-elicitation-complete
     text: 'Server returns complete result after retry with inputResponses (elicitation)'
 
-  - check: input-required-result-sampling-complete
+  - check: sep-2322-sampling-complete
     text: 'Server returns complete result after retry with inputResponses (sampling)'
 
-  - check: input-required-result-list-roots-complete
+  - check: sep-2322-list-roots-complete
     text: 'Server returns complete result after retry with inputResponses (list roots)'
 
-  - check: input-required-result-request-state-complete
+  - check: sep-2322-request-state-complete
     text: 'Server returns complete result after retry with requestState echoed back'
 
-  - check: input-required-result-multiple-inputs-complete
+  - check: sep-2322-multiple-inputs-complete
     text: 'Server returns complete result after retry with multiple inputResponses'
 
   # ── Multi-Round Flow ───────────────────────────────────────────────────────
-  - check: input-required-result-multi-round-r1
+  - check: sep-2322-multi-round-r1
     text: 'Server returns InputRequiredResult on first round of multi-round flow'
 
-  - check: input-required-result-multi-round-r2
+  - check: sep-2322-multi-round-r2
     text: 'Server returns InputRequiredResult on second round (re-requesting with new requestState)'
 
-  - check: input-required-result-multi-round-r3
+  - check: sep-2322-multi-round-r3
     text: 'Server returns complete result on third round of multi-round flow'
 
   # ── Non-Tool Request (prompts/get) ─────────────────────────────────────────
-  - check: input-required-result-non-tool-incomplete
+  - check: sep-2322-non-tool-incomplete
     text: 'Server returns InputRequiredResult on prompts/get request'
 
-  - check: input-required-result-non-tool-complete
+  - check: sep-2322-non-tool-complete
     text: 'Server returns complete GetPromptResult after retry with inputResponses'
 
   # ── Error Handling ─────────────────────────────────────────────────────────
@@ -92,7 +92,7 @@ requirements:
   - check: sep-2322-ignore-unexpected-params
     text: 'If additional, unexpected parameters are provided in the InputResponses object, the server SHOULD ignore any information it does not recognize or need.'
 
-  - check: input-required-result-missing-response-rerequests
+  - check: sep-2322-missing-response-rerequests
     text: 'If the client fails to send all the information requested in a previous InputRequests, and the missing information is necessary for the server to process the request, the server SHOULD respond with a new InputRequiredResult requesting the missing information again, rather than returning an error.'
 
   # ── Excluded requirements ──────────────────────────────────────────────────
